+++ src/main/java/cart/domain/cart/service/CartService.java
@@ -1,10 +1,10 @@
 package cart.domain.cart.service;
 
-import cart.domain.cart.dao.CartDao;
+import cart.dao.CartDao;
 import cart.domain.cart.entity.Cart;
-import cart.domain.member.dao.MemberDao;
+import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
-import cart.domain.product.dao.ProductDao;
+import cart.dao.ProductDao;
 import cart.domain.product.entity.Product;
 import cart.dto.CartCreateRequest;
 import cart.dto.CartCreateResponse;
