+++ src/main/java/cart/domain/product/service/ProductService.java
@@ -1,6 +1,6 @@
 package cart.domain.product.service;
 
-import cart.domain.product.dao.ProductDao;
+import cart.dao.ProductDao;
 import cart.domain.product.entity.Product;
 import cart.dto.ProductCreateRequest;
 import cart.dto.ProductResponse;
