+++ src/test/java/cart/domain/cart/service/CartServiceTest.java
@@ -7,17 +7,16 @@ import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 
-import cart.domain.cart.dao.CartDao;
+import cart.dao.CartDao;
 import cart.domain.cart.entity.Cart;
-import cart.domain.member.dao.MemberDao;
+import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
-import cart.domain.product.dao.ProductDao;
+import cart.dao.ProductDao;
 import cart.domain.product.entity.Product;
 import cart.dto.CartCreateRequest;
 import cart.dto.CartCreateResponse;
 import java.time.LocalDateTime;
 import java.util.Optional;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
