+++ src/test/java/cart/domain/member/dao/MemberDaoTest.java
@@ -2,6 +2,7 @@ package cart.domain.member.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import cart.dao.MemberDao;
 import cart.domain.member.entity.Member;
 import java.time.LocalDateTime;
 import java.util.List;
